refactor(All code): Change project structure
Sort files to folders to make it easily understandable
