+++ ipfilter/ipfilter.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <iostream>
+
+void ipfilter(std::istream& sin, std::ostream& sout);
