--- lib.cpp
@@ -1,7 +1,0 @@
-#include "lib.h"
-
-#include "version.h"
-
-int version() {
-	return PROJECT_VERSION_PATCH;
-}
