+++ lib/lib.cpp
@@ -1,0 +1,5 @@
+#include "lib.h"
+
+#include "version.h"
+
+int version() { return PROJECT_VERSION_PATCH; }
