+++ lib/version.h.in
